+++ app/api/hello/route.ts
@@ -1,0 +1,3 @@
+export async function GET(request: Request) {
+  return new Response('Hello, Next.js!')
+}
